frontend/requirements.txt
streamlit>=1.31.0
requests>=2.31.0
sqlparse>=0.4.4
Pygments>=2.17.2
python-dotenv>=1.0.0 